Completed ADX Library.
ADX Library is now complete, you can pretty much change about anything when encoding provided the correct WAV file.
With the exception of BitDepth, I am not planning to add this, maybe in the far future, but not even Cri's own ADX lib support multiple bitdepth. It's hard to implement and moreover, going under 4 bits will present noticeable data loss, and going over that will defeat the point of compression. The max in can go is 16 anyway. So the options are from 1 - 15.