Improved HCA encoding and decoding with Encryption
Made an improvised port of VGAudio, and also added in the better HCA encrypting and decrypting methods, now they work in C.
I also changed the PCM code a bit, as well as fixing bugs in the IO handlers.

I do have an improved ADX code ready, but would need me to test it to push it eventually. TODO for now.